folder naming correction
