+++ Plans-shop/Projects/plantsshop.website/src/App.jsx
@@ -2,9 +2,9 @@ import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { NavBar } from './components/NavBar/NavBar';
 import Footer from './components/Footer/Footer';
-import Home from './pages/Home';
-import Contact from './pages/Contact';
-import AboutUs from './pages/AboutUs';
+import Home from './pages/HomePage/Home';
+import Contact from './pages/ContactPage/Contact';
+import AboutUs from './pages/AboutUsPage/AboutUs';
 import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
