--- Plans-shop/Projects/plantsshop.website/src/pages/Contact.jsx
@@ -1,15 +1,0 @@
-import FAQ from "./ContactComponents/FAQSection/FAQ";
-import Form from "./ContactComponents/FormSection/Form";
-import ContactHero from "./ContactComponents/HeroSection/ContactHero";
-
-export default function Contact() {
-  return (
-    <div className="home">
-      <ContactHero />
-      <Form />
-      <FAQ />
-
-      
-    </div>
-  );
-}
