+++ Plans-shop/Projects/plantsshop.website/src/pages/ContactPage/Contact.jsx
@@ -1,0 +1,15 @@
+import FAQ from "./FAQSection/FAQ";
+import Form from "./FormSection/Form";
+import ContactHero from "./HeroSection/ContactHero";
+
+export default function Contact() {
+  return (
+    <div className="home">
+      <ContactHero />
+      <Form />
+      <FAQ />
+
+      
+    </div>
+  );
+}
