+++ Plans-shop/Projects/plantsshop.website/src/pages/ContactPage/HeroSection/ContactHero.jsx
@@ -1,6 +1,6 @@
 import { Typography } from "@mui/material";
 import benefitsBig from "../../../assets/benefitsBig.jpg";
-import "../../AboutUsComponents/HeroSection/AboutUsHero.css";
+import "../../AboutUsPage/HeroSection/AboutUsHero.css";
 export default function ContactHero() {
   return (
     <div className="aboutus-section">
