+++ Plans-shop/Projects/plantsshop.website/src/pages/Shop.jsx
@@ -23,7 +23,7 @@ import FavoriteOutlinedIcon from "@mui/icons-material/FavoriteOutlined";
 import FavoriteBorderOutlinedIcon from "@mui/icons-material/FavoriteBorderOutlined";
 import ShoppingCartOutlinedIcon from "@mui/icons-material/ShoppingCartOutlined";
 import AddShoppingCartOutlinedIcon from "@mui/icons-material/AddShoppingCartOutlined";
-import "./HomeComponents/ArrivalsSection/Arrivals.css";
+import "./HomePage/ArrivalsSection/Arrivals.css";
 
 const ITEMS_PER_PAGE = 9;
